Add adding S invoices to T invoice

+++ app/src/main/java/kz/kazpost/loadingarea/repositories/s_invoice/models/AddSInvoicesToTInvoiceRequest.kt
@@ -7,7 +7,7 @@ data class AddSInvoicesToTInvoiceRequest(
     val sInvoiceIdsSeparatedByComma: String,
 
     @field:SerializedName("tlId")
-    val tInvoiceId: Long,
+    val tInvoiceId: Int,
 
     @field:SerializedName("averageVolume")
     val averageVolume: Long,
